+++ src/components/TextEditor/extensions/image-group/ImageGroupUploadDialog.vue
@@ -1,14 +1,12 @@
 <template>
   <Dialog
     v-model="modelValue"
-    :options="{
-      title: props.mode === 'edit' ? 'Edit Images' : 'Upload Images',
-      size: '3xl',
-    }"
+    :title="props.mode === 'edit' ? 'Edit Images' : 'Upload Images'"
+    size="3xl"
     @close="$emit('close')"
-    :disableOutsideClickToClose="true"
+    :dismissible="false"
   >
-    <template #body-content>
+    <template #default>
       <div class="space-y-2">
         <div class="flex items-center gap-2">
           <Button @click="triggerFileInput">
@@ -230,7 +228,7 @@
       </div>
     </template>
   </Dialog>
-  <Teleport to="body">
+  <Teleport :to="portalTarget ?? 'body'">
     <Transition
       name="fade"
       enter-active-class="transition-opacity duration-200"
@@ -253,6 +251,7 @@
 </template>
 
 <script setup lang="ts">
+import { usePortalTarget } from '../../../../composables/usePortalTarget'
 import {
   computed,
   ref,
@@ -302,6 +301,8 @@ const props = withDefaults(
     mode: 'new',
   },
 )
+
+const portalTarget = usePortalTarget()
 
 const emit = defineEmits(['update:modelValue', 'close', 'update:files', 'save'])